Improve HTTP request parsing and pass cookies to CGI scripts
The HTTP request parser has been mostly rewritten, and now it is able to
parse header lines. This makes it possible to get the cookie string from
the request and pass it to CGI scripts, which was needed for the cflip
forum to retain login status.

It is now possible to run the forum through cfws, log in and submit
threads and posts!